# docs/swagger.yaml
basePath: /api
definitions:
  handlers.MailResponse:
    properties:
      account_type:
        type: string
      content_type:
        type: string
      created_at:
        type: string
      html:
        description: Gedecodeerde/gesanitized HTML body
        type: string
      id:
        type: string
      message_id:
        type: string
      processed_at:
        type: string
      read:
        type: boolean
      received_at:
        type: string
      sender:
        type: string
      subject:
        type: string
      to:
        type: string
      uid:
        type: string
      updated_at:
        type: string
    type: object
  handlers.PaginatedMailResponse:
    properties:
      emails:
        items:
          $ref: '#/definitions/handlers.MailResponse'
        type: array
      totalCount:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact:
    email: info@dekoninklijkeloop.nl
    name: API Support
    url: http://www.dekoninklijkeloop.nl
  description: API voor de De Koninklijke Loop email service - inclusief registratie,
    deelnemersbeheer, CMS en meer
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: DKL Email Service API
  version: "1.0"
paths:
...
  /api/mail:
    get:
      consumes:
      - application/json
      description: Haalt een gepagineerde lijst van emails op
      parameters:
      - description: Aantal resultaten per pagina (standaard 10)
        in: query
        name: limit
        type: integer
      - description: Offset voor paginering (standaard 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PaginatedMailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Lijst van emails ophalen
      tags:
      - Mail
  /api/mail/{id}:
    delete:
      consumes:
      - application/json
      description: Verwijdert een email uit het systeem
      parameters:
      - description: Email ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Email verwijderen
      tags:
      - Mail
    get:
      consumes:
      - application/json
      description: Haalt de details van een specifieke email op
      parameters:
      - description: Email ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Details van een specifieke email ophalen
      tags:
      - Mail
  /api/mail/{id}/processed:
    put:
      consumes:
      - application/json
      description: Markeert een email als verwerkt om aan te geven dat deze is afgehandeld
      parameters:
      - description: Email ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.MailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Email als verwerkt markeren
      tags:
      - Mail
  /api/mail/account/{type}:
    get:
      consumes:
      - application/json
      description: Haalt een gepagineerde lijst van emails op voor een specifiek account
        type
      parameters:
      - description: Account Type (bv. 'info', 'inschrijving')
        in: path
        name: type
        required: true
        type: string
      - description: Aantal resultaten per pagina (standaard 10)
        in: query
        name: limit
        type: integer
      - description: Offset voor paginering (standaard 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PaginatedMailResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Lijst van emails ophalen per account type
      tags:
      - Mail
  /api/mail/fetch:
    post:
      consumes:
      - application/json
      description: Haalt nieuwe emails op van de mailserver en slaat deze op in de
        database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Nieuwe emails ophalen
      tags:
      - Mail
  /api/mail/unprocessed:
    get:
      consumes:
      - application/json
      description: Haalt een lijst van alle onverwerkte emails op
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PaginatedMailResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Onverwerkte emails ophalen
      tags:
      - Mail
...
securityDefinitions:
  BearerAuth:
    description: JWT Bearer token voor geauthenticeerde endpoints
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"